Fix bad move bug.
This bug was introduced when I added timing interrupts to negamax.
It seems to choose a bad move if it runs out of time in the middle
of the while loop. If I allow it to evaluate every state at depth
X and then check the time, it will pick a good move. Will work on
adding interrupts while maintaining good move choices later.